+++ slog/internal/middleware/auth/middleware.go
@@ -32,7 +32,7 @@ func WithAuth(authorizer Authorizer, next http.HandlerFunc) http.HandlerFunc {
 		}
 
 		// Set authorized information to context
-		ctx := logger.WithAuthorizedInfo(r.Context(), *authInfo)
+		ctx := logger.SetAuthorizedInfoCtx(r.Context(), *authInfo)
 
 		// Update request context
 		r = r.WithContext(ctx)
